Fixed infinite loading bug, prepped for eventsystem
Fixed the rotochair bug that caused unity to run forever when pressing play. Had to move the streaming assets folder to fix it.

Moved some stuff around in RotoManager to prepare for the eventsystem's addition. There's now some redundant functions in Rotomanager but once we transition fully to the eventsystem, the redundant functions will be deleted.